Begin Cpu/Chipeit classes draft
