--- src/chipeitlib/src/main/java/es/chipeit/lib/interfaces/ICpu.kt
@@ -1,2 +1,0 @@
-package es.chipeit.lib.interfaces
-
